remove unneeded parentheses around binary shifts
Shifts have higher precedence then binary and/or/xor. While shifts are
like multiplication/division, 'and' and 'or' are similar to
subtraction/addition, so the precedence is easy to remember - if not:
'man operator' always helps (-;

Some arguments have been reordered to make the evaluation order (from
left to right more apparent.