introduce the 'hval' data type for heap values
Start to distinguish between the width of the heap cells and the width
of the values they contain (related to #85). For now, 'hval' should be
used for everything non-pointer. The nice side effect is, that
calculations regarding immediates and headers are done in 32 bit by
default (leading to shorter x86_64 code), which makes masking with FMAX
unnecessary - this should have been done inside F() anyway.